Reorganize making the directions
Now, instead of doing all the work for the directions of a hallway in Hallway.getDirectionsFromIndices, we call methods from the Room and Turn classes in when we're passing, entering, and leaving those rooms. That means that we can customize the directions for passing, entering, and leaving individual rooms by overriding those methods.
This shouldn't cause any change in behavior.