Turn off autocomplete
So that the browser doesn't try to autocomplete your input based on what you've previously searched (on desktop browsers).
It's also helpful on mobile browsers because it discourages the browser from trying to autocomplete based on words instead of based on the datalist.
I previously thought that this would cause the datalist to stop working in the Opera browser, but it seems like it works.